fix: compatibilidad en el paquete Marca, se elimina DTOs innecesarios y se ajusta controlador y Mapper.

+++ src/main/java/com/coagronet/marca/dtos/MarcaDTO.java
@@ -11,6 +11,6 @@ public class MarcaDTO {
     private Long id;
     private String nombre;
     private String descripcion;
-    private Integer estado;
+    private Long estado;
     private Long empresa;
 }
